delete sub

+++ be/src/main/java/com/netcracker/edu/backend/repository/SubscriptionRepository.java
@@ -7,6 +7,6 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface SubscriptionRepository extends JpaRepository<Subscription, Long> {
+public interface SubscriptionRepository extends JpaRepository<Subscription, Integer> {
     List<Subscription> findByUserId(Integer id);
 }
